+++ app/src/main/java/com/neqabty/domain/entities/NotificationEntity.kt
@@ -16,7 +16,7 @@ data class NotificationEntity(
     var name: String?,
     var trip: String?,
     var regiment: String?,
-    var cost: Int?,
+    var cost: Double?,
     var housingType: String?,
     var numChild: String?,
     var phone: String?
